requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cu111
torchvision==0.10.0+cu111
ConfigArgParse==1.5.3
einops==0.3.0
GPUtil==1.4.0
nvidia-smi==0.1.3
nvidia-ml-py3==7.352.0
omegaconf==2.2.3
open3d==0.9.0.0
opencv-python==4.5.1.48
Pillow==8.4.0
plotly==4.14.3
plyfile==0.7.4
trimesh==3.15.5
tqdm
# numpy==1.19.5
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
tensorflow==1.15.0
hydra-core==1.2.0
loguru==0.6.0
pyhocon==0.3.59
shapely==1.8.5.post1 # for trimesh